Fix LOA steps display on special authorizations tab, remove LOA stepper and vehicle section from LOA steps, update LOAFormData
