--- frontend/src/features/settings/components/SpecialAuthorizations/LOA/vehicles/VehicleTable.scss
@@ -1,46 +1,0 @@
-@import "../../../../../../themes/orbcStyles";
-
-.loa-vehicle-table {
-  color: $bc-black;
-
-  &.table-container {
-    padding-bottom: 0;
-    margin-top: 0;
-  }
-
-  &__search {
-    .search-label {
-      font-weight: bold;
-      margin-bottom: 0.5rem;
-
-      &--light {
-        font-weight: normal;
-      }
-    }
-
-    .search-input {
-      &__input-container {
-        border: 2px solid $bc-text-box-border-grey;
-
-        fieldset {
-          border: none;
-        }
-      }
-    }
-  }
-
-  & &__container {
-    border: 1px solid $bc-border-grey;
-    width: auto;
-  }
-
-  & &__paper-container {
-    border-radius: 0;
-  }
-
-  & &__checkbox {
-    &--error {
-      color: $bc-red;
-    }
-  }
-}
