+++ frontend/src/features/settings/pages/SpecialAuthorizations/LOA/LOASteps.tsx
@@ -1,19 +1,21 @@
 import { useContext, useEffect, useMemo, useState } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faArrowLeft, faArrowRight } from "@fortawesome/free-solid-svg-icons";
-import { Button, Step, StepConnector, StepLabel, Stepper } from "@mui/material";
+import { faPencil } from "@fortawesome/free-solid-svg-icons";
+import { Button } from "@mui/material";
 import { FormProvider, useForm } from "react-hook-form";
 import { useNavigate } from "react-router-dom";
 
 import "./LOASteps.scss";
 import { LOAStep, LOA_STEPS, labelForLOAStep } from "../../../types/LOAStep";
-import { LOADesignateVehicles } from "./vehicles/LOADesignateVehicles";
 import { LOAReview } from "./review/LOAReview";
 import { LOABasicInfo } from "./basic/LOABasicInfo";
 import { Nullable } from "../../../../../common/types/common";
 import { LOAFormData, loaDetailToFormData } from "../../../types/LOAFormData";
 import { ERROR_ROUTES } from "../../../../../routes/constants";
 import { SnackBarContext } from "../../../../../App";
+import { usePowerUnitSubTypesQuery } from "../../../../manageVehicles/hooks/powerUnits";
+import { useTrailerSubTypesQuery } from "../../../../manageVehicles/hooks/trailers";
+import { getDefaultRequiredVal } from "../../../../../common/helpers/util";
 import {
   useCreateLOAMutation,
   useFetchLOADetail,
@@ -30,12 +32,6 @@ export const LOASteps = ({
   companyId: number;
   onExit: () => void;
 }) => {
-  const steps = [
-    labelForLOAStep(LOA_STEPS.BASIC),
-    labelForLOAStep(LOA_STEPS.VEHICLES),
-    labelForLOAStep(LOA_STEPS.REVIEW),
-  ];
-
   const navigate = useNavigate();
   const { setSnackBar } = useContext(SnackBarContext);
   const { data: loaDetail } = useFetchLOADetail(companyId, loaId);
@@ -43,6 +39,18 @@ export const LOASteps = ({
   const updateLOAMutation = useUpdateLOAMutation();
   const removeLOADocumentMutation = useRemoveLOADocumentMutation();
   const loaFormData = loaDetailToFormData(loaDetail);
+
+  const powerUnitSubTypesQuery = usePowerUnitSubTypesQuery();
+  const trailerSubTypesQuery = useTrailerSubTypesQuery();
+  const powerUnitSubtypes = useMemo(() => getDefaultRequiredVal(
+    [],
+    powerUnitSubTypesQuery.data,
+  ), [powerUnitSubTypesQuery.data]);
+
+  const trailerSubtypes = useMemo(() => getDefaultRequiredVal(
+    [],
+    trailerSubTypesQuery.data,
+  ), [trailerSubTypesQuery.data]);
 
   const formMethods = useForm<LOAFormData>({
     defaultValues: loaFormData,
@@ -56,25 +64,19 @@ export const LOASteps = ({
   }, [loaDetail]);
 
   const [activeStep, setActiveStep] = useState<LOAStep>(LOA_STEPS.BASIC);
-
-  const showPrevBtn =
-    activeStep === LOA_STEPS.VEHICLES || activeStep === LOA_STEPS.REVIEW;
-  const showNextBtn =
-    activeStep === LOA_STEPS.BASIC || activeStep === LOA_STEPS.VEHICLES;
+  const stepLabel = labelForLOAStep(activeStep, loaId);
+  const showPrevBtn = activeStep === LOA_STEPS.REVIEW;
+  const showNextBtn = activeStep === LOA_STEPS.BASIC;
   const showFinishBtn = activeStep === LOA_STEPS.REVIEW;
 
   const handlePrev = () => {
-    if (activeStep === LOA_STEPS.VEHICLES) {
+    if (activeStep === LOA_STEPS.REVIEW) {
       setActiveStep(LOA_STEPS.BASIC);
-    } else if (activeStep === LOA_STEPS.REVIEW) {
-      setActiveStep(LOA_STEPS.VEHICLES);
     }
   };
 
   const handleNext = () => {
     if (activeStep === LOA_STEPS.BASIC) {
-      setActiveStep(LOA_STEPS.VEHICLES);
-    } else if (activeStep === LOA_STEPS.VEHICLES) {
       setActiveStep(LOA_STEPS.REVIEW);
     }
   };
@@ -125,56 +127,30 @@ export const LOASteps = ({
 
   const stepComponent = useMemo(() => {
     switch (activeStep) {
-      case LOA_STEPS.VEHICLES:
-        return <LOADesignateVehicles />;
       case LOA_STEPS.REVIEW:
-        return <LOAReview />;
+        return (
+          <LOAReview
+            powerUnitSubtypes={powerUnitSubtypes}
+            trailerSubtypes={trailerSubtypes}
+          />
+        );
       default:
-        return <LOABasicInfo onRemoveDocument={handleRemoveDocument} />;
+        return (
+          <LOABasicInfo
+            onRemoveDocument={handleRemoveDocument}
+            powerUnitSubtypes={powerUnitSubtypes}
+            trailerSubtypes={trailerSubtypes}
+          />
+        );
     }
   }, [activeStep]);
 
   return (
     <FormProvider {...formMethods}>
       <div className="loa-steps">
-        <Stepper
-          className="stepper"
-          activeStep={activeStep}
-          alternativeLabel
-          connector={
-            <StepConnector
-              className="step__connector"
-              classes={{ line: "step__connector-line" }}
-            />
-          }
-        >
-          {steps.map((label, stepIndex) => (
-            <Step
-              className={`step ${stepIndex === 0 ? "step--first" : ""}`}
-              key={label}
-            >
-              <StepLabel
-                className="step__label"
-                classes={{
-                  labelContainer: "step__label-container",
-                  active: "step__label--active",
-                  disabled: "step__label--disabled",
-                  completed: "step__label--completed",
-                }}
-                StepIconProps={{
-                  className: "step__icon",
-                  classes: {
-                    text: "step__step-number",
-                    active: "step__icon--active",
-                    completed: "step__icon--completed",
-                  },
-                }}
-              >
-                {label}
-              </StepLabel>
-            </Step>
-          ))}
-        </Stepper>
+        <h3 className="loa-steps__step-label">
+          {stepLabel}
+        </h3>
 
         <div className="loa-steps__step-component">{stepComponent}</div>
 
@@ -198,10 +174,10 @@ export const LOASteps = ({
               onClick={handlePrev}
               variant="contained"
               color="secondary"
-              startIcon={<FontAwesomeIcon icon={faArrowLeft} />}
+              startIcon={<FontAwesomeIcon icon={faPencil} />}
               className="steps-navigation__btn steps-navigation__btn--prev"
             >
-              <strong>Previous</strong>
+              <strong>Edit</strong>
             </Button>
           ) : null}
 
@@ -213,9 +189,8 @@ export const LOASteps = ({
               onClick={handleSubmit(handleNext)}
               variant="contained"
               color="primary"
-              endIcon={<FontAwesomeIcon icon={faArrowRight} />}
             >
-              Next
+              Continue
             </Button>
           ) : null}
