+++ frontend/src/features/settings/pages/SpecialAuthorizations/LOA/review/LOAReview.scss
@@ -14,8 +14,4 @@
   &__header {
     font-weight: bold;
   }
-
-  .loa-vehicle-table {
-    margin-top: 0.5rem;
-  }
 }
