--- frontend/src/features/settings/pages/SpecialAuthorizations/LOA/vehicles/LOADesignateVehicles.scss
@@ -1,22 +1,0 @@
-@import "../../../../../../themes/orbcStyles";
-
-.loa-designate-vehicles {
-  color: $bc-black;
-
-  &__header {
-    font-weight: bold;
-    font-size: 1.25rem;
-  }
-
-  &__info-banner {
-    margin: 1.5rem 0;
-  }
-
-  &__info-message {
-    font-weight: normal;
-  }
-
-  &__selection-error {
-    color: $bc-red;
-  }
-}
