+++ frontend/src/features/settings/types/LOAFormData.ts
@@ -2,7 +2,8 @@ import { Dayjs } from "dayjs";
 
 import { Nullable } from "../../../common/types/common";
 import { PERMIT_TYPES } from "../../permits/types/PermitType";
-import { LOAVehicle } from "./LOAVehicle";
+import { VehicleType } from "../../manageVehicles/types/Vehicle";
+import { DEFAULT_VEHICLE_TYPE } from "../../manageVehicles/types/Vehicle";
 import { LOADetail } from "./LOADetail";
 import {
   applyWhenNotNullable,
@@ -34,10 +35,8 @@ export interface LOAFormData {
     fileName: string;
   }> | File;
   additionalNotes?: Nullable<string>;
-  selectedVehicles: {
-    powerUnits: Record<string, Nullable<LOAVehicle>>;
-    trailers: Record<string, Nullable<LOAVehicle>>;
-  };
+  vehicleType: VehicleType;
+  vehicleSubtype: string;
 }
 
 /**
@@ -63,23 +62,18 @@ export const loaDetailToFormData = (
     loaDetail?.startDate,
     now(),
   );
+
   const expiryDate = applyWhenNotNullable(
     expiryDateStr => getEndOfDate(toLocalDayjs(expiryDateStr)),
     loaDetail?.expiryDate,
     null,
   );
+
   const neverExpires = !expiryDate;
   const additionalNotes = getDefaultRequiredVal("", loaDetail?.comment);
-  const powerUnits = applyWhenNotNullable(
-    powerUnitsArr => Object.fromEntries(powerUnitsArr.map(powerUnitId => [powerUnitId, null])),
-    loaDetail?.powerUnits,
-    {},
-  );
-  const trailers = applyWhenNotNullable(
-    trailersArr => Object.fromEntries(trailersArr.map(trailerId => [trailerId, null])),
-    loaDetail?.trailers,
-    {},
-  );
+  const vehicleType = getDefaultRequiredVal(DEFAULT_VEHICLE_TYPE, loaDetail?.vehicleType);
+  const vehicleSubtype = getDefaultRequiredVal("", loaDetail?.vehicleSubtype);
+
   const defaultFile = loaDetail?.documentId ? {
     fileName: getDefaultRequiredVal("", loaDetail?.fileName),
   } : null;
@@ -91,10 +85,8 @@ export const loaDetailToFormData = (
     neverExpires,
     uploadFile: defaultFile,
     additionalNotes,
-    selectedVehicles: {
-      powerUnits,
-      trailers,
-    },
+    vehicleType,
+    vehicleSubtype,
   };
 };
 
@@ -112,9 +104,6 @@ export const serializeLOAFormData = (loaFormData: LOAFormData) => {
     })
     .map(([permitType]) => permitType);
 
-  const powerUnits = Object.keys(loaFormData.selectedVehicles.powerUnits);
-  const trailers = Object.keys(loaFormData.selectedVehicles.trailers);
-  
   const body = {
     startDate: dayjsToLocalStr(loaFormData.startDate, DATE_FORMATS.DATEONLY),
     expiryDate: loaFormData.expiryDate
@@ -122,8 +111,8 @@ export const serializeLOAFormData = (loaFormData: LOAFormData) => {
       : null,
     comment: getDefaultRequiredVal("", loaFormData.additionalNotes),
     loaPermitType: permitTypes,
-    powerUnits: powerUnits.length > 0 ? powerUnits : undefined,
-    trailers: trailers.length > 0 ? trailers : undefined,
+    vehicleType: loaFormData.vehicleType,
+    vehicleSubtype: loaFormData.vehicleSubtype,
   };
 
   if (loaFormData.uploadFile instanceof File) {
